+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(inlpubs)
-
-test_check("inlpubs")