+++ NEWS.md
@@ -1,6 +1,15 @@
+# inlpubs 1.0.4
+
+- Add additional citations
+- Remove dependency on **inlmisc** package.
+
+# inlpubs 1.0.3
+
+- Configure package website for GitHub deployment.
+
 # inlpubs 1.0.2
 
-- Remove dependency on **inlpubs** package in inst/CITATION file.
+- Remove dependency on **inlpubs** package in the CITATION file.
 
 # inlpubs 1.0.1